refactor: doc_index to index
