--- docarray/doc_index/__init__.py
@@ -1,3 +1,0 @@
-from docarray.doc_index.backends.hnswlib import HnswDocumentIndex
-
-__all__ = ['HnswDocumentIndex']
